example/requirements.txt
anyio==3.5.0; python_version >= "3.6" and python_full_version >= "3.6.2"
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.6"
fastapi==0.73.0; python_full_version >= "3.6.1"
fastapi-cloud-logging==0.8.4; python_version >= "3.7"
google-api-core==2.7.2; python_version >= "3.6"
google-auth==2.6.5; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
google-cloud-appengine-logging==1.1.1; python_version >= "3.6"
google-cloud-audit-log==0.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
google-cloud-core==2.3.0; python_version >= "3.6"
google-cloud-logging==3.0.0; python_version >= "3.6"
googleapis-common-protos==1.56.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
grpc-google-iam-v1==0.12.4; python_version >= "3.6"
grpcio-status==1.45.0; python_version >= "3.6"
grpcio==1.45.0; python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6")
gunicorn==20.1.0; python_version >= "3.6" 
idna==3.3; python_version >= "3.6" and python_full_version >= "3.6.2" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
packaging==21.3; python_version >= "3.6"
proto-plus==1.20.3; python_version >= "3.6"
protobuf==3.20.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") or python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
pydantic==1.9.0; python_full_version >= "3.6.1"
pyparsing==3.0.8; python_full_version >= "3.6.8" and python_version >= "3.6"
requests==2.27.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6") or python_full_version >= "3.6.0" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6")
sniffio==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.2"
starlette==0.17.1; python_version >= "3.6" and python_full_version >= "3.6.1"
typing-extensions==4.1.1; python_version >= "3.6" and python_full_version >= "3.6.1"
urllib3==1.26.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
uvicorn[standard]==0.17.6; python_version >= "3.6"